restructuring update
