+++ pyproject.toml
@@ -9,7 +9,11 @@ python = "^3.8"
 ariadne = "^0.12.0"
 starlette = "^0.13.0"
 uvicorn = "^0.12.2"
+starlette-jwt = "^0.1.9"
+pandas = "^1.1.4"
 python-dateutil = "^2.8.1"
+aiomysql = "^0.0.20"
+boolean = "^1.1.0"
 pre-commit = "^2.8.2"
 
 [tool.poetry.dev-dependencies]
