--- saltapi/auth/__init__.py
@@ -1,1 +1,0 @@
-"""Authentication and authorization."""
