+++ saltapi/repository/user_repository.py
@@ -2,39 +2,11 @@ from typing import Optional, Dict
 
 import pandas as pd
 
-from saltapi.database.sdb_connection import sdb_connect
+from saltapi.repository.database import sdb_connect
 from saltapi.util.error import InvalidUsage
 
 
-def query_user_id(username: str) -> Optional[int]:
-    """
-    Query the PIPT user id.
-
-    Parameters
-    ----------
-    username : str
-        The PIPT username.
-
-    Returns
-    -------
-    user_is: int
-        The PIPT user id.
-    """
-    sql = """SELECT PiptUser_Id
-             FROM PiptUser
-             WHERE Username='{username}'""" \
-        .format(username=username)
-
-    conn = sdb_connect()
-    try:
-        result = pd.read_sql(sql, conn)
-        conn.close()
-        return result.iloc[0]['PiptUser_Id']
-    except IndexError:
-        return None
-
-
-def verify_user(username: str, password: str) -> Optional:
+async def find_user_id_by_credentials(username: str, password: str) -> Optional:
     """
     Verify the username and password of the user.
 
@@ -48,21 +20,21 @@ def verify_user(username: str, password: str) -> Optional:
     -------
     None.
     """
-    sql = """SELECT COUNT(PiptUser_Id) AS UserCount
+    sql = """SELECT PiptUser_Id
              FROM PiptUser
              WHERE Username='{username}' AND Password=MD5('{password}')""" \
         .format(username=username, password=password)
 
-    conn = sdb_connect()
+    conn = await sdb_connect()
     result = pd.read_sql(sql, conn)
     conn.close()
-    if not result.iloc[0]['UserCount']:
+    if not result.iloc[0]['PiptUser_Id']:
         raise InvalidUsage('Username or password wrong')
 
-    return True
+    return result.iloc[0]['PiptUser_Id']
 
 
-def query_user(user_id: int) -> Dict:
+async def find_user_by_id(user_id: int) -> Dict:
     """
     Query user details
     Parameters
@@ -79,7 +51,7 @@ def query_user(user_id: int) -> Dict:
         JOIN Investigator AS i using (Investigator_Id)
     WHERE u.PiptUser_Id = {user_id}
     '''.format(user_id=user_id)
-    conn = sdb_connect()
+    conn = await sdb_connect()
     results = pd.read_sql(sql, conn)
     conn.close()
